+++ frontend/app/src/components/history/events/dialog-types.ts
@@ -12,6 +12,7 @@ export const DIALOG_TYPES = {
   ADD_TRANSACTION: 'addTransaction',
   DECODING_STATUS: 'decodingStatus',
   EVENT_FORM: 'eventForm',
+  MATCH_ASSET_MOVEMENTS: 'matchAssetMovements',
   MISSING_RULES: 'missingRules',
   PROTOCOL_CACHE: 'protocolCache',
   REPULLING_TRANSACTION: 'repullingTransaction',
@@ -23,6 +24,7 @@ export type DialogType = typeof DIALOG_TYPES[keyof typeof DIALOG_TYPES];
 export type DialogShowOptions =
   | { type: typeof DIALOG_TYPES.ADD_MISSING_RULE; data: AccountingRuleIdentifier }
   | { type: typeof DIALOG_TYPES.EVENT_FORM; data: GroupEventData | StandaloneEventData }
+  | { type: typeof DIALOG_TYPES.MATCH_ASSET_MOVEMENTS }
   | { type: typeof DIALOG_TYPES.TRANSACTION_FORM; data?: AddTransactionHashPayload }
   | { type: typeof DIALOG_TYPES.REPULLING_TRANSACTION }
   | { type: typeof DIALOG_TYPES.MISSING_RULES; data: HistoryEventEditData }