+++ frontend/app/src/components/settings/GeneralSettingsCategory.vue
@@ -8,6 +8,7 @@ import CsvExportDelimiterSetting from '@/components/settings/general/CsvExportDe
 import DateDisplayFormatSetting from '@/components/settings/general/DateDisplayFormatSetting.vue';
 import DateInputFormatSetting from '@/components/settings/general/DateInputFormatSetting.vue';
 import DisplayDateInLocaltimeSetting from '@/components/settings/general/DisplayDateInLocaltimeSetting.vue';
+import EventsProcessingFrequencySetting from '@/components/settings/general/EventsProcessingFrequencySetting.vue';
 import UsageAnalyticsSetting from '@/components/settings/general/UsageAnalyticsSetting.vue';
 import VersionUpdateFrequencySetting from '@/components/settings/general/VersionUpdateFrequencySetting.vue';
 import SettingCategory from '@/components/settings/SettingCategory.vue';
@@ -37,6 +38,7 @@ const { t } = useI18n({ useScope: 'global' });
     <AskUserUponSizeDiscrepancySetting />
     <VersionUpdateFrequencySetting />
     <BalanceSaveFrequencySetting />
+    <EventsProcessingFrequencySetting />
     <BtcDerivationGapLimitSetting />
     <SettingsItem>
       <template #title>